Further restructuring
Moved and renamed libidlc files to their own folder
Moved and renamed generator_common files to their own folder
Moved idlc files to their own folder
