+++ src/tools/idlc/include/generator_common/common.h
@@ -8,11 +8,19 @@
 //
 // SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
 
-#ifndef IDL_DEFS_H
-#define IDL_DEFS_H
+#ifndef GENERATOR_COMMON_COMMON_H
+#define GENERATOR_COMMON_COMMON_H
 
-#include <stdbool.h>
+#include "common_export.h"
+
+#include <assert.h>
 #include <stdint.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <inttypes.h>
+#include <stdbool.h>
+
 #include "idl/processor.h"
 
 #if defined (__cplusplus)
@@ -39,7 +47,7 @@ struct idlc_option {
   char *help;
 };
 
-typedef struct {
+struct idlc_generator_config {
   char *output_dir; /* path to write completed files */
   char* base_dir; /* Path to start reconstruction of dir structure */
 
@@ -52,14 +60,21 @@ typedef struct {
   Passing a pointer to a generator function is a reasonable way of avoiding the layering problems
   this introduces. May be a null pointer */
   idl_retcode_t (*generate_typeinfo_typemap) (const idl_pstate_t *pstate, const idl_node_t *node, idl_typeinfo_typemap_t *result);
-} idlc_generator_config_t;
+};
+typedef struct idlc_generator_config idlc_generator_config_t;
 
 typedef const idlc_option_t **(*idlc_generator_options_t)(void);
 typedef const idl_builtin_annotation_t **(*idlc_generator_annotations_t)(void);
 typedef int(*idlc_generate_t)(const idl_pstate_t *, const idlc_generator_config_t *);
 
+IDL_COMMON_EXPORT int
+print_type(char *str, size_t size, const void *ptr, void *user_data);
+
+IDL_COMMON_EXPORT int
+print_scoped_name(char *str, size_t size, const void *ptr, void *user_data);
+
 #if defined(__cplusplus)
 }
 #endif
 
-#endif /* IDL_DEFS_H */
+#endif /* GENERATOR_COMMON_COMMON_H */
