+++ src/tools/idlc/src/generator_common/descriptor_type_meta.c
@@ -34,7 +34,6 @@
 #include "idl/string.h"
 #include "idl/misc.h"
 #include "descriptor_type_meta.h"
-#include "generator.h"
 
 static struct dds_cdrstream_allocator idlc_cdrstream_default_allocator = { idl_malloc, idl_realloc, idl_free };
