+++ src/tools/idlc/src/idlc/idlc.c
@@ -32,10 +32,12 @@
 #include "mcpp_lib.h"
 #include "mcpp_out.h"
 
-#include "libidlc/idl_defs.h"
 #include "plugin.h"
 #include "options.h"
-#include "libidlc/descriptor_type_meta.h"
+#include "common.h"
+#ifdef DDS_HAS_TYPELIB
+#include "descriptor_type_meta.h"
+#endif
 #include "file.h"
 
 #if 0
