+++ src/tools/idlc/src/idlc/plugin.c
