+++ src/tools/idlc/src/libidlc/libidlc__types.h
@@ -11,7 +11,7 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-#include "libidlc/idl_defs.h"
+#include "generator_common/common.h"
 
 idl_retcode_t generate_types(const idl_pstate_t *pstate, struct generator *generator);
