+++ src/tools/idlc/tests/test_common.h
@@ -12,7 +12,7 @@
 #define IDLC_TEST_COMMON_H
 
 #include <stdlib.h>
-#include "descriptor.h"
+#include "libidlc/libidlc__descriptor.h"
 
 idl_retcode_t generate_test_descriptor (idl_pstate_t *pstate, const char *idl, struct descriptor *descriptor);
